+++ package.json
@@ -40,6 +40,7 @@
     ]
   },
   "devDependencies": {
+    "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.27.1",
